--- core/src/main/scala/za/co/absa/spline/harvester/dispatcher/modelmapper/UntypedMapConverter.scala
@@ -1,34 +1,0 @@
-/*
- * Copyright 2023 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.spline.harvester.dispatcher.modelmapper
-
-import za.co.absa.spline.producer.model.AttrOrExprRef
-
-class UntypedMapConverter(toAttrOrExprRef: (AttrOrExprRef => Any)) {
-
-  def toUntypedMap(map: Map[String, Any]): Map[String, Any] = {
-    def mapValue(value: Any): Any = value match {
-      case aoe: AttrOrExprRef => toAttrOrExprRef(aoe)
-      case s: Seq[_] => s.map(mapValue)
-      case m: Map[_, _] => m.transform((k, v) => mapValue(v))
-      case o: Option[_] => o.map(mapValue)
-      case other => other
-    }
-
-    map.transform((k, v) => mapValue(v))
-  }
-}
